requirements.txt
python-telegram-bot==20.6
anthropic==0.9.0
python-dotenv==1.0.0
gitpython==3.1.36
fastapi==0.103.1
uvicorn==0.23.2
langchain==0.0.267
requests==2.31.0
pydantic==2.3.0
pillow==10.0.0
pyautogui==0.9.54
